refactor: useDeploy hook
